Move manual test cases owned by tpapaioa out of test_manual.py
